requirements.txt
# Dépendances de base
pandas>=1.5.0
python-dotenv>=0.19.0

# PostgreSQL
psycopg2-binary>=2.9.0

# Firebase
firebase-admin>=6.2.0

# Logging et utilitaires
typing-extensions>=4.0.0
